+++ app/src/main/java/nextstep/jwp/WebServer.java
@@ -1,5 +1,6 @@
 package nextstep.jwp;
 
+import nextstep.jwp.infrastructure.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
